refactor: update paths to work for this PR but will break reorganizing components commits
