+++ src/common/helpers/ComponentExampleBase.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as styles from './ComponentExampleBase.scss';
 import IReactComponentProps from '../structures/IReactComponentProps';
 import Header from '../../components/Header';
-import Button from '../../components/Button';
+import { Button } from '../../components/buttons/Button/Button';
 
 export interface IComponentExampleBasePropDetails {
 	defaultValue?: any;
