+++ src/components/FlyLargeConfirm/FlyLargeConfirm.tsx
@@ -4,7 +4,7 @@ import * as styles from './FlyLargeConfirm.sass';
 import CheckSVG from '../../svg/checkmark--big';
 import IReactComponentProps from '../../common/structures/IReactComponentProps';
 import Handler from '../../common/structures/Handler';
-import Button from '../Button';
+import { PrimaryButton } from '../buttons/PrimaryButton/PrimaryButton';
 
 interface IProps extends IReactComponentProps {
 
@@ -52,8 +52,7 @@ export default class FlyLargeConfirm extends React.Component<IProps, IState> {
 			>
 				<p>{this.props.bodyText}</p>
 
-				<Button
-					recipe="primary"
+				<PrimaryButton
 					onClick={this.onConfirm}
 					disabled={this.state.confirmed}
 				>
@@ -61,7 +60,7 @@ export default class FlyLargeConfirm extends React.Component<IProps, IState> {
 						<CheckSVG />
 					)}
 					{this.state.confirmed ? 'Confirmed' : this.props.buttonText}
-				</Button>
+				</PrimaryButton>
 			</div>
 		);
 	}
