+++ src/components/VerticalNav/components/WorkspaceSwitcher.tsx
@@ -4,7 +4,6 @@ import * as styles from './WorkspaceSwitcher.scss';
 import * as stylesVerticalNav from '../VerticalNav.scss';
 import Popup from '../../Popup/Popup';
 import Divider from '../../Divider/Divider';
-import Button from '../../Button/Button';
 import AddSVG from '../../../svg/add';
 import Avatar from '../../Avatar';
 import IReactComponentProps from '../../../common/structures/IReactComponentProps';
@@ -12,6 +11,7 @@ import Handler from '../../../common/structures/Handler';
 import { VerticalNavItem } from '../VerticalNav';
 import CloseSmallSVG from '../../../svg/close--small';
 import WarningSVG from '../../../svg/warning';
+import { TextButton } from '../../buttons/TextButton/TextButton';
 
 interface IWorkspaceSwitcherProps extends IReactComponentProps {
 	className?: string;
@@ -182,12 +182,9 @@ export class WorkspaceSwitcher extends React.Component<IWorkspaceSwitcherProps, 
 							<Divider />
 								<>
 									<div className={styles.WorkspaceSwitcher_Section}>
-										<Button
-											recipe="text"
-											onClick={this.props.onClickUpgradeToPro}
-										>
+										<TextButton onClick={this.props.onClickUpgradeToPro}>
 											UPGRADE TO PRO
-										</Button>
+										</TextButton>
 									</div>
 									<Divider />
 								</>
@@ -197,8 +194,7 @@ export class WorkspaceSwitcher extends React.Component<IWorkspaceSwitcherProps, 
 									styles.WorkspaceSwitcher_Section,
 								)}
 							>
-								<Button
-									recipe="text"
+								<TextButton
 									onClick={() => (
 										this.state.activeWorkspaceItem.isOwner
 											?
@@ -208,13 +204,10 @@ export class WorkspaceSwitcher extends React.Component<IWorkspaceSwitcherProps, 
 									)}
 								>
 									{this.state.activeWorkspaceItem.isOwner ? 'Manage Team' : 'My Account'}
-								</Button>
-								<Button
-									recipe="text"
-									onClick={this.props.onClickLogout}
-								>
+								</TextButton>
+								<TextButton onClick={this.props.onClickLogout}>
 									Logout
-								</Button>
+								</TextButton>
 							</div>
 						</Popup>
 						:
